chore: table data move types to types.ts
