+++ apps/web/src/app/[lang]/app/[type]/parties/table-data.tsx
@@ -1,19 +1,12 @@
 import type {
-  UniRefund_CRMService_Customss_CreateCustomsDto,
+  UniRefund_CRMService_AddressTypes_CreateAddressTypeWithComponentsDto,
   UniRefund_CRMService_Customss_CreateCustomsOrganizationDto,
-  UniRefund_CRMService_Merchants_CreateMerchantDto,
+  UniRefund_CRMService_EmailCommonDatas_CreateEmailCommonDataWithComponentsDto,
   UniRefund_CRMService_Merchants_CreateMerchantOrgnaizationDto,
-  UniRefund_CRMService_RefundPoints_CreateRefundPointDto,
   UniRefund_CRMService_RefundPoints_CreateRefundPointOrganizationDto,
-  UniRefund_CRMService_TaxFrees_CreateTaxFreeDto,
   UniRefund_CRMService_TaxFrees_CreateTaxFreeOrganizationDto,
-  UniRefund_CRMService_TaxOffices_CreateTaxOfficeDto,
   UniRefund_CRMService_TaxOffices_CreateTaxOfficeOrganizationDto,
-  Volo_Abp_Application_Dtos_PagedResultDto_110,
-  Volo_Abp_Application_Dtos_PagedResultDto_13,
-  Volo_Abp_Application_Dtos_PagedResultDto_16,
-  Volo_Abp_Application_Dtos_PagedResultDto_17,
-  Volo_Abp_Application_Dtos_PagedResultDto_19,
+  UniRefund_CRMService_TelephoneTypes_CreateTelephoneTypeWithComponentsDto,
 } from "@ayasofyazilim/saas/CRMService";
 import {
   $UniRefund_CRMService_Customss_CustomsProfileDto,
@@ -27,49 +20,9 @@ import {
   $UniRefund_CRMService_TaxFrees_CreateTaxFreeDto as CreateTaxFreeSchema,
   $UniRefund_CRMService_TaxOffices_CreateTaxOfficeDto as CreateTaxOfficeSchema,
 } from "@ayasofyazilim/saas/CRMService";
-import type {
-  addressTypeCodes,
-  emailTypeCodes,
-  telephoneTypeCodes,
-} from "@repo/ui/utils/table/form-schemas";
 import { ContactFormSubPositions } from "@repo/ui/utils/table/form-schemas";
 import { PhoneNumberUtil } from "google-libphonenumber";
-
-export type CreateMerchantDTO =
-  UniRefund_CRMService_Merchants_CreateMerchantDto;
-export type CreateCustomsDTO = UniRefund_CRMService_Customss_CreateCustomsDto;
-export type CreateRefundPointDTO =
-  UniRefund_CRMService_RefundPoints_CreateRefundPointDto;
-export type CreateTaxFreeDTO = UniRefund_CRMService_TaxFrees_CreateTaxFreeDto;
-export type CreateTaxOfficeDTO =
-  UniRefund_CRMService_TaxOffices_CreateTaxOfficeDto;
-
-export type GetMerchantDTO = Volo_Abp_Application_Dtos_PagedResultDto_16;
-export type GetRefundPointDTO = Volo_Abp_Application_Dtos_PagedResultDto_17;
-export type GetCustomsDTO = Volo_Abp_Application_Dtos_PagedResultDto_13;
-export type GetTaxFreeDTO = Volo_Abp_Application_Dtos_PagedResultDto_19;
-export type GetTaxOfficeDTO = Volo_Abp_Application_Dtos_PagedResultDto_110;
-
-export type PartyNameType =
-  | "merchants"
-  | "refund-points"
-  | "customs"
-  | "tax-free"
-  | "tax-offices";
-
-export type PartiesResultType =
-  | GetMerchantDTO
-  | GetRefundPointDTO
-  | GetCustomsDTO
-  | GetTaxFreeDTO
-  | GetTaxOfficeDTO;
-
-export type PartiesCreateDTOType =
-  | CreateMerchantDTO
-  | CreateCustomsDTO
-  | CreateRefundPointDTO
-  | CreateTaxFreeDTO
-  | CreateTaxOfficeDTO;
+import type { PartiesCreateType } from "./types";
 
 export interface CreatePartiesDto {
   taxOfficeId: string;
@@ -79,36 +32,10 @@ export interface CreatePartiesDto {
     | UniRefund_CRMService_RefundPoints_CreateRefundPointOrganizationDto
     | UniRefund_CRMService_TaxFrees_CreateTaxFreeOrganizationDto
     | UniRefund_CRMService_TaxOffices_CreateTaxOfficeOrganizationDto;
-  telephone: {
-    areaCode: string;
-    localNumber: string;
-    ituCountryCode: string;
-    primaryFlag: boolean;
-    typeCode: telephoneTypeCodes;
-  };
-  address: {
-    addressLine: string;
-    city: string;
-    terriority: string;
-    postalCode: string;
-    country: string;
-    fullAddress: string;
-    primaryFlag: boolean;
-    typeCode: addressTypeCodes;
-  };
-  email: {
-    emailAddress: string;
-    primaryFlag: boolean;
-    typeCode: emailTypeCodes;
-  };
+  telephone: UniRefund_CRMService_TelephoneTypes_CreateTelephoneTypeWithComponentsDto;
+  address: UniRefund_CRMService_AddressTypes_CreateAddressTypeWithComponentsDto;
+  email: UniRefund_CRMService_EmailCommonDatas_CreateEmailCommonDataWithComponentsDto;
 }
-
-export type PartiesCreateType =
-  | typeof CreateMerchantSchema
-  | typeof CreateRefundPointSchema
-  | typeof CreateCustomsSchema
-  | typeof CreateTaxFreeSchema
-  | typeof CreateTaxOfficeSchema;
 
 const CommonOrganizationFields = ["name", "taxpayerId", "branchId"];
 const OrganizationFields = ["customerNumber", "legalStatusCode"];
